perf(bundle): lazy-load recharts and react-grid-layout
Wrap recharts and react-grid-layout imports with React.lazy() and
Suspense boundaries to keep them out of the initial bundle. Extract
BotPnlChart and TradingGrid into dedicated feature modules. Consolidate
scattered mock data into src/lib/mock-data.ts.

Initial bundle reduced from 263KB to 141KB gzipped (<=200KB PRD target).

Closes #20
