+++ src/features/portfolio/balance-display.test.tsx
@@ -1,29 +1,42 @@
 import { render, screen } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
+import { MOCK_PORTFOLIO_STATE } from "@/lib/mock-data";
 import { BalanceDisplay } from "./balance-display";
 
 describe("BalanceDisplay", () => {
   it("renders total balance", () => {
     render(
-      <BalanceDisplay totalBalance={10000.0} availableBalance={5000.0} unrealizedPnL={1500.0} />,
+      <BalanceDisplay
+        totalBalance={MOCK_PORTFOLIO_STATE.totalBalance}
+        availableBalance={MOCK_PORTFOLIO_STATE.availableBalance}
+        unrealizedPnL={MOCK_PORTFOLIO_STATE.unrealizedPnL}
+      />,
     );
     expect(screen.getByText(/Total/)).toBeInTheDocument();
-    expect(screen.getByText(/10000\.00/)).toBeInTheDocument();
+    expect(screen.getByText(/10,?423\.70/)).toBeInTheDocument();
   });
 
   it("renders available balance", () => {
     render(
-      <BalanceDisplay totalBalance={10000.0} availableBalance={5000.0} unrealizedPnL={1500.0} />,
+      <BalanceDisplay
+        totalBalance={MOCK_PORTFOLIO_STATE.totalBalance}
+        availableBalance={MOCK_PORTFOLIO_STATE.availableBalance}
+        unrealizedPnL={MOCK_PORTFOLIO_STATE.unrealizedPnL}
+      />,
     );
     expect(screen.getByText(/Available/)).toBeInTheDocument();
-    expect(screen.getByText(/5000\.00/)).toBeInTheDocument();
+    expect(screen.getByText(/4,?328\.50/)).toBeInTheDocument();
   });
 
   it("renders positive PnL with profit color", () => {
     render(
-      <BalanceDisplay totalBalance={10000.0} availableBalance={5000.0} unrealizedPnL={1500.0} />,
+      <BalanceDisplay
+        totalBalance={MOCK_PORTFOLIO_STATE.totalBalance}
+        availableBalance={MOCK_PORTFOLIO_STATE.availableBalance}
+        unrealizedPnL={MOCK_PORTFOLIO_STATE.unrealizedPnL}
+      />,
     );
-    const pnlElement = screen.getByText(/1500\.00/);
+    const pnlElement = screen.getByText(/416\.10/);
     expect(pnlElement).toHaveClass("text-[color:var(--trading-profit)]");
   });
