+++ src/routes/bots.$botId.tsx
@@ -1,10 +1,10 @@
 import { createFileRoute, Link } from '@tanstack/react-router'
 import { ErrorBoundary } from '@/ui/error-boundary'
-import { useState } from 'react'
+import { lazy, Suspense, useState } from 'react'
+
+const BotPnlChart = lazy(() => import('@/features/bots/bot-pnl-chart'))
 import { ArrowLeft, Pencil, X, Check, Play, Pause, Square } from 'lucide-react'
-import { LineChart, Line, ResponsiveContainer, Tooltip, XAxis, YAxis, ReferenceLine } from 'recharts'
-import { MOCK_BOTS } from '@/features/bots/mock-bots'
-import { MOCK_BOT_TRADES } from '@/features/bots/mock-bot-trades'
+import { MOCK_BOTS, MOCK_BOT_TRADES } from '@/lib/mock-data'
 import type { BotStrategy, BotStatus } from '@/features/bots/types'
 
 export const Route = createFileRoute('/bots/$botId')({
@@ -147,47 +147,9 @@ function BotDetailPage() {
           {/* P&L Chart — 2/3 */}
           <div className="lg:col-span-2 rounded-md border border-border/60 p-4">
             <p className="text-[10px] text-muted-foreground uppercase tracking-wide mb-3">P&L History</p>
-            <ResponsiveContainer width="100%" height={200}>
-              <LineChart
-                data={bot.pnlHistory.map((v, i) => ({ tick: i + 1, pnl: v }))}
-                margin={{ top: 4, right: 4, bottom: 0, left: 0 }}
-              >
-                <XAxis
-                  dataKey="tick"
-                  tick={{ fontSize: 9, fill: 'var(--color-muted-foreground)' }}
-                  axisLine={false}
-                  tickLine={false}
-                  tickFormatter={(v) => `T${v}`}
-                />
-                <YAxis
-                  tick={{ fontSize: 9, fill: 'var(--color-muted-foreground)' }}
-                  axisLine={false}
-                  tickLine={false}
-                  tickFormatter={(v) => (v >= 0 ? `+${v}` : String(v))}
-                  width={36}
-                />
-                <ReferenceLine y={0} stroke="var(--color-border)" strokeDasharray="3 3" />
-                <Tooltip
-                  contentStyle={{
-                    background: 'var(--color-card)',
-                    border: '1px solid var(--color-border)',
-                    borderRadius: 6,
-                    fontSize: 11,
-                    color: 'var(--color-foreground)',
-                  }}
-                  formatter={(value) => { const n = Number(value); return [(n >= 0 ? '+' : '') + n.toFixed(2), 'P&L'] }}
-                  labelFormatter={(label) => `Trade ${label}`}
-                />
-                <Line
-                  type="monotone"
-                  dataKey="pnl"
-                  stroke={totalPnl >= 0 ? 'var(--trading-profit)' : 'var(--trading-loss)'}
-                  strokeWidth={1.5}
-                  dot={false}
-                  activeDot={{ r: 3 }}
-                />
-              </LineChart>
-            </ResponsiveContainer>
+            <Suspense fallback={<div className="w-full h-[200px] rounded bg-muted animate-pulse" />}>
+              <BotPnlChart pnlHistory={bot.pnlHistory} totalPnl={totalPnl} />
+            </Suspense>
           </div>
 
           {/* Settings — 1/3 */}
